Remove check on source.client in irc_join handler.
If the user wasn't already known to us by being in another channel, then
m->source.client will be null despite bot->find_client succeeding. This
is because the message structure was created before the handler ran that
made the client in question known.

This should really be fixed by allowing the handlers in channel.cpp to
modify the message structure, but that's a longer piece of work.