Add support for RPL_QUIETLIST(728) and RPL_ENDOFQUIETLIST(729) and fix memory leak in banlist handling
Adds support for the new quiet numerics introduced in Charybdis 3.2.
For backwards compatibility the overloading of RPL_BANLIST and RPL_ENDOFBANLIST is still supported.

This patch also fixes an issue in banlist handling which would cause duplicate entries to be
stored if another script or module requests banlists from the server.