Don't lowercase hostmask when checking against admin list
Right now, the user's hostmask is lowercased according to RFC 1459 conventions before it is checked against the admins regex. This is undesirable because it means the regex needs to be explicitly written with that implementation detail in mind. The regex cannot simply be passed a case-insensitive flag because RFC casing also transforms []\ into {}|. As such, it is preferable to remove the requirement to know an eir implementation detail at the expense of requiring proper casing in the admins regex.